# book-network-ui/src/app/services/models/book-request.ts
/* tslint:disable */
/* eslint-disable */
export interface BookRequest {
  authorName: string;
  id?: number;
  isbn: string;
  shareable?: boolean;
  synopsis: string;
  title: string;
  genre?: string; // Add genre field
}